Alterado para Long o atributo do endpoint de consulta de estados por id. Liberado o recurso estados para consulta nas configurações de segurança

+++ src/main/java/br/com/sinaldasorte/resource/EstadoResource.java
@@ -35,7 +35,7 @@ public class EstadoResource {
 	}
 	
 	@RequestMapping(value="/{estadoId}/cidades", method=RequestMethod.GET)
-	public ResponseEntity<List<CidadeDTO>> findCidades(@PathVariable Integer estadoId) {
+	public ResponseEntity<List<CidadeDTO>> findCidades(@PathVariable Long estadoId) {
 		List<Cidade> list = cidadeService.findByEstado(estadoId);
 		List<CidadeDTO> listDto = list.stream().map(obj -> new CidadeDTO(obj)).collect(Collectors.toList());  
 		return ResponseEntity.ok().body(listDto);
