+++ src/main/java/br/com/sinaldasorte/service/CidadeService.java
@@ -14,7 +14,7 @@ public class CidadeService {
 	@Autowired
 	private CidadeRepository repo;
 
-	public List<Cidade> findByEstado(Integer estadoId) {
+	public List<Cidade> findByEstado(Long estadoId) {
 		return repo.findCidades(estadoId);
 	}
 }
